pyproject.toml
[project]
name = "chat-gpt-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.2.0",
]
